Refactor into slots
